+++ libs/platform/src/lib/components/form/form-options.ts
@@ -5,3 +5,7 @@ export type HintPlacement = 'left' | 'right';
 export type FormZone = 'zTop' | 'zBottom' | 'zLeft' | 'zRight';
 
 export type Column = 1 | 2 | 3 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
+
+export type ContentDensity = 'compact' | 'cozy';
+
+export type Status = 'success' | 'error' | 'warning' | 'default' | 'information';
