+++ libs/platform/src/lib/components/form/radio-group/radio-group.component.ts
@@ -25,7 +25,7 @@ import { Subject } from 'rxjs';
 import { RadioButtonComponent } from './radio/radio.component';
 import { CollectionBaseInput } from '../collection-base.input';
 import { FormFieldControl } from '../form-control';
-import { FormField } from '../form-group/form-field/form-field';
+import { FormField } from '../form-field';
 
 /**
  * Radio group implementation based on the
@@ -55,7 +55,7 @@ export class RadioGroupComponent extends CollectionBaseInput implements AfterVie
     }
 
     /**
-     * To Dispaly Radio buttons in a line
+     * To Display Radio buttons in a line
      */
     @Input()
     isInline = false;
@@ -111,14 +111,14 @@ export class RadioGroupComponent extends CollectionBaseInput implements AfterVie
     }
 
     /**
-     * Acess display value for objects, acts as checkbox label.
+     * Access display value for objects, acts as checkbox label.
      */
     public getDisplayValue(item: any): string {
         return this.displayValue(item);
     }
 
     /**
-     * Acess lookup value for objects, acts as checkbox value.
+     * Access lookup value for objects, acts as checkbox value.
      */
     public getLookupValue(item: any): string {
         return this.lookupValue(item);
@@ -231,7 +231,7 @@ export class RadioGroupComponent extends CollectionBaseInput implements AfterVie
 
     /**
      *
-     * @param button Set inital values, used while content children creation
+     * @param button Set initial values, used while content children creation
      */
     private _setProperties(button: RadioButtonComponent): void {
         if (button) {
